Moved chartutils to chartbasics + added to Van Krevelen Diagram
