+++ src/main/java/net/sf/mzmine/chartbasics/chartutils/XYBlockPixelSizePaintScales.java
@@ -16,7 +16,7 @@
  * USA
  */
 
-package net.sf.mzmine.modules.visualization.kendrickmassplot.chartutils;
+package net.sf.mzmine.chartbasics.chartutils;
 
 import java.awt.Color;
 import java.awt.Paint;
